requirements.txt
asgiref==3.5.2
astroid==2.13.3
beautifulsoup4==4.11.1
dill==0.3.6
Django==4.1.1
django-bootstrap-static==4.2.1
django-bootstrap5==22.1
django-cities==0.6.1
django-countries==7.4.2
django-debug-toolbar==3.7.0
django-guardian==2.4.0
django-model-utils==4.2.0
factory-boy==3.2.1
Faker==17.3.0
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.3.0
platformdirs==2.6.2
pylint==2.15.10
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
swapper==1.3.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
wrapt==1.14.1
